chore: invitation workflow
