--- frontend/src/pages/Login2.jsx
@@ -1,267 +1,0 @@
-import { useState, useEffect } from "react";
-import { motion } from "framer-motion";
-import { Mail, Lock, ArrowRight } from "lucide-react";
-import { Link, useNavigate, useSearchParams } from "react-router-dom";
-import PasswordField from "../components/PasswordField";
-import Button from "../components/Button";
-import { authService } from "../services/authService";
-import { useRedirectIfLoggedIn } from "../hooks/useRedirectIfLoggedIn";
-
-export default function Login() {
-  const navigate = useNavigate();
-  const [searchParams] = useSearchParams();
-  const redirectUrl = searchParams.get("redirect");
-
-  useRedirectIfLoggedIn(redirectUrl);
-
-  const [formData, setFormData] = useState({
-    email: "",
-    password: "",
-  });
-  const [errors, setErrors] = useState({});
-  const [touched, setTouched] = useState({});
-
-  const validate = () => {
-    const newErrors = {};
-
-    if (!formData.email) {
-      newErrors.email = "Email is required";
-    } else if (!/\S+@\S+\.\S+/.test(formData.email)) {
-      newErrors.email = "Email is invalid";
-    }
-
-    if (!formData.password) {
-      newErrors.password = "Password is required";
-    } else if (formData.password.length < 6) {
-      newErrors.password = "Password must be at least 6 characters";
-    }
-
-    setErrors(newErrors);
-    return Object.keys(newErrors).length === 0;
-  };
-
-  const handleChange = (e) => {
-    const { name, value } = e.target;
-    setFormData({
-      ...formData,
-      [name]: value,
-    });
-
-    if (errors[name]) {
-      setErrors({
-        ...errors,
-        [name]: "",
-      });
-    }
-  };
-
-  const handleBlur = (e) => {
-    setTouched({
-      ...touched,
-      [e.target.name]: true,
-    });
-  };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    setTouched({
-      email: true,
-      password: true,
-    });
-
-    const isValid = validate();
-    if (isValid) {
-      try {
-        localStorage.removeItem("access");
-        const response = await authService.login(formData);
-        localStorage.setItem("access", response.data.access);
-
-        // Redirect to the specified URL or dashboard
-        if (redirectUrl) {
-          navigate(redirectUrl);
-        } else {
-          navigate("/dashboard");
-        }
-      } catch (error) {
-        setErrors({
-          general: "Invalid email or password. Please try again.",
-        });
-      }
-    }
-  };
-
-  return (
-    <div className="min-h-screen bg-gradient-to-br from-indigo-100 via-white to-purple-100 flex flex-col items-center justify-center px-6">
-      {/* Navbar */}
-      <nav className="w-full flex justify-between items-center py-6 max-w-6xl absolute top-0">
-        <img src="/src/assets/splitzy.svg" className="mr-3 h-12" alt="Logo" />
-      </nav>
-
-      {/* Login Form */}
-      <motion.div
-        initial={{ opacity: 0, y: 40 }}
-        animate={{ opacity: 1, y: 0 }}
-        transition={{ duration: 0.6 }}
-        className="w-full max-w-md"
-      >
-        <div className="bg-white shadow-xl rounded-2xl p-8">
-          <div className="text-center mb-8">
-            <h2 className="text-3xl font-bold text-gray-800 mb-2">
-              Welcome Back
-            </h2>
-            <p className="text-gray-600">Log in to manage your expenses</p>
-            {redirectUrl && (
-              <p className="text-sm text-indigo-600 mt-2">
-                Please log in to continue
-              </p>
-            )}
-          </div>
-
-          <div className="space-y-5">
-            {/* General Error */}
-            {errors.general && (
-              <div className="bg-red-50 border border-red-200 rounded-lg p-3">
-                <p className="text-sm text-red-600">{errors.general}</p>
-              </div>
-            )}
-
-            {/* Email Input */}
-            <div>
-              <label
-                htmlFor="email"
-                className="block text-sm font-medium text-gray-700 mb-2"
-              >
-                Email Address
-              </label>
-              <div className="relative">
-                <Mail className="absolute left-3 top-1/2 transform -translate-y-1/2 h-5 w-5 text-gray-400" />
-                <input
-                  type="email"
-                  id="email"
-                  name="email"
-                  value={formData.email}
-                  onChange={handleChange}
-                  onBlur={handleBlur}
-                  placeholder="Enter your email"
-                  className={`w-full pl-11 pr-4 py-3 border rounded-lg focus:outline-none focus:ring-2 focus:ring-indigo-600 focus:border-transparent transition ${
-                    touched.email && errors.email
-                      ? "border-red-500"
-                      : "border-gray-300"
-                  }`}
-                />
-              </div>
-              {touched.email && errors.email && (
-                <p className="mt-1 text-sm text-red-600">{errors.email}</p>
-              )}
-            </div>
-
-            {/* Password Input */}
-            <div>
-              <label
-                htmlFor="password"
-                className="block text-sm font-medium text-gray-700 mb-2"
-              >
-                Password
-              </label>
-              <div className="relative">
-                <Lock className="absolute left-3 top-1/2 transform -translate-y-1/2 h-5 w-5 text-gray-400" />
-                <PasswordField
-                  password={formData.password}
-                  handleChange={handleChange}
-                  onBlur={handleBlur}
-                  name="password"
-                  className={`w-full pl-11 pr-4 py-3 border rounded-lg focus:outline-none focus:ring-2 focus:ring-indigo-600 focus:border-transparent transition ${
-                    touched.password && errors.password
-                      ? "border-red-500"
-                      : "border-gray-300"
-                  }`}
-                />
-              </div>
-              {touched.password && errors.password && (
-                <p className="mt-1 text-sm text-red-600">{errors.password}</p>
-              )}
-            </div>
-
-            {/* Forgot Password */}
-            <div className="flex justify-end m-2">
-              <Link
-                to="#"
-                className="text-sm text-indigo-600 hover:text-indigo-700 font-medium"
-              >
-                Forgot password?
-              </Link>
-            </div>
-
-            {/* Submit Button */}
-            <Button
-              onClick={handleSubmit}
-              className="w-full bg-indigo-600 text-white py-3 rounded-lg font-medium hover:bg-indigo-700 transition-all duration-200 flex items-center justify-center gap-2"
-            >
-              Log In
-              <ArrowRight className="h-5 w-5" />
-            </Button>
-          </div>
-
-          {/* Divider */}
-          <div className="mt-6 text-center text-sm text-gray-600">
-            Don't have an account?{" "}
-            <Link
-              to={
-                redirectUrl
-                  ? `/signup?redirect=${encodeURIComponent(redirectUrl)}`
-                  : "/signup"
-              }
-              className="text-indigo-600 hover:text-indigo-700 font-medium"
-            >
-              Sign up
-            </Link>
-          </div>
-        </div>
-
-        {/* Alternative Login Options */}
-        <div className="mt-6">
-          <div className="relative">
-            <div className="absolute inset-0 flex items-center">
-              <div className="w-full border-t border-gray-300"></div>
-            </div>
-            <div className="relative flex justify-center text-sm">
-              <span className="px-4 bg-gradient-to-br from-indigo-100 via-white to-purple-100 text-gray-500">
-                Or continue with
-              </span>
-            </div>
-          </div>
-
-          <div className="mt-6 grid grid-cols-2 gap-3">
-            <button className="w-full bg-white border border-gray-300 py-3 rounded-lg font-medium hover:bg-gray-50 transition-all duration-200 flex items-center justify-center gap-2">
-              <svg className="h-5 w-5" viewBox="0 0 24 24">
-                <path
-                  fill="#4285F4"
-                  d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"
-                />
-                <path
-                  fill="#34A853"
-                  d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"
-                />
-                <path
-                  fill="#FBBC05"
-                  d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"
-                />
-                <path
-                  fill="#EA4335"
-                  d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"
-                />
-              </svg>
-              Google
-            </button>
-            <button className="w-full bg-white border border-gray-300 py-3 rounded-lg font-medium hover:bg-gray-50 transition-all duration-200 flex items-center justify-center gap-2">
-              <svg className="h-5 w-5" fill="#1877F2" viewBox="0 0 24 24">
-                <path d="M24 12.073c0-6.627-5.373-12-12-12s-12 5.373-12 12c0 5.99 4.388 10.954 10.125 11.854v-8.385H7.078v-3.47h3.047V9.43c0-3.007 1.792-4.669 4.533-4.669 1.312 0 2.686.235 2.686.235v2.953H15.83c-1.491 0-1.956.925-1.956 1.874v2.25h3.328l-.532 3.47h-2.796v8.385C19.612 23.027 24 18.062 24 12.073z" />
-              </svg>
-              Facebook
-            </button>
-          </div>
-        </div>
-      </motion.div>
-    </div>
-  );
-}
